Issue #4587: Moved resources for avoidescapedunicodecharacters
