+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/AvoidEscapedUnicodeCharactersCheckTest.java
@@ -145,7 +145,10 @@ public class AvoidEscapedUnicodeCharactersCheckTest extends BaseCheckTestSupport
 
     @Override
     protected String getPath(String filename) throws IOException {
-        return super.getPath("checks" + File.separator + filename);
+        return super.getPath("checks" + File.separator
+                + "misc" + File.separator
+                + "avoidescapedunicodecharacters" + File.separator
+                + filename);
     }
 
     @Test
