+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/avoidescapedunicodecharacters/InputAvoidEscapedUnicodeCharacters.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks;
+package com.puppycrawl.tools.checkstyle.checks.misc.avoidescapedunicodecharacters;
 
 import java.util.concurrent.TimeUnit;
